% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmCOS.R
\name{fitLM}
\alias{fitLM}
\title{Fit a Bayesian linear model}
\usage{
fitLM(y, X, Z, wts, priors, n.samples = 1000)
}
\arguments{
\item{y}{A numeric vector of response values.}

\item{X}{A numeric matrix of predictors (design matrix).}

\item{Z}{A vector consisting of a covariate.}

\item{wts}{A numeric vector of weights for the observations.}

\item{priors}{A list containing prior distributions for the model parameters.}

\item{n.samples}{Number of posterior samples to draw.}
}
\value{
An object of class \code{fitLM} containing the fitted model and posterior samples.
}
\description{
Fit a Bayesian linear model
}
