% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmCOS.R
\name{modularLM}
\alias{modularLM}
\title{Fit a Bayesian linear model under modular framework for COS}
\usage{
modularLM(y, X, Z, wts, priors)
}
\arguments{
\item{y}{A numeric vector of response values.}

\item{X}{A numeric matrix of predictors (design matrix).}

\item{Z}{A numeric matrix containg posterior samples of a covariate (e.g., spatial-temporal process).
Each column corresponds to a different sample, and number of rows must match the length of \code{y}.}

\item{wts}{A numeric vector of weights for the observations.}

\item{priors}{A list containing prior distributions for the model parameters.}
}
\value{
An object of class \code{modularLM} containing the fitted model and posterior samples.
}
\description{
Fit a Bayesian linear model under modular framework for COS
}
