% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptPredictCOS.R
\name{sptBlockPredictTimeAgg}
\alias{sptBlockPredictTimeAgg}
\title{Block-level spatial-temporal prediction}
\usage{
sptBlockPredictTimeAgg(
  mod.out,
  Xcov_pred,
  polycoords_pred,
  timecoords_pred,
  nMCpoly_coords = 500,
  verbose = TRUE
)
}
\arguments{
\item{mod.out}{An object of class \code{sptLMexactCOSTimeAgg} containing the fitted model.}

\item{Xcov_pred}{A numeric matrix of basis functions for the prediction.}

\item{polycoords_pred}{An \code{sf} object containing the polygons for upscale prediction.}

\item{timecoords_pred}{A numeric matrix of time intervals for upscale prediction.}

\item{nMCpoly_coords}{An integer specifying the number of Monte Carlo points to sample inside each polygon.}

\item{verbose}{A logical value indicating whether to print progress messages.}
}
\value{
An object of class \code{sptBlockPredictTimeAgg} containing the posterior predictive samples.
}
\description{
Block-level spatial-temporal prediction
}
