% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptLMexact.R
\name{sptLMexactTimeAgg}
\alias{sptLMexactTimeAgg}
\title{Univariate Bayesian spatial-temporal linear model for misaligned data}
\usage{
sptLMexactTimeAgg(
  X,
  Xcov,
  spcoords,
  timecoords,
  phi_s,
  phi_t,
  nu_s,
  noise_sp_ratio,
  n.samples = 100,
  verbose = TRUE
)
}
\arguments{
\item{X}{A numeric vector of values (e.g., exposure).}

\item{Xcov}{A numeric matrix of covariates (e.g., basis functions).}

\item{spcoords}{A numeric matrix of spatial coordinates (longitude, latitude).}

\item{timecoords}{A numeric matrix of time intervals (e.g., year, month).}

\item{phi_s}{A numeric value for the spatial decay parameter.}

\item{phi_t}{A numeric value for the temporal decay parameter.}

\item{nu_s}{A numeric value for the smoothness parameter.}

\item{noise_sp_ratio}{A numeric value for the noise to spatial-temporal variance ratio.}

\item{n.samples}{An integer specifying the number of posterior samples to draw.}

\item{verbose}{A logical value indicating whether to print progress messages.}
}
\value{
An object of class \code{sptLMexactCOSTimeAgg} containing the fitted model and posterior samples.
}
\description{
Univariate Bayesian spatial-temporal linear model for misaligned data
}
