% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptPredictCOS.R
\name{sptPointPredictTimeAgg}
\alias{sptPointPredictTimeAgg}
\title{Point-level spatial-temporal prediction}
\usage{
sptPointPredictTimeAgg(mod.out, Xcov_pred, spcoords_pred, timecoords_pred)
}
\arguments{
\item{mod.out}{An object of class \code{sptLMexactCOSTimeAgg} containing the fitted model.}

\item{Xcov_pred}{A numeric matrix of basis functions for the prediction.}

\item{spcoords_pred}{A numeric matrix of spatial coordinates for downscale prediction (longitude, latitude).}

\item{timecoords_pred}{A numeric matrix of temporal coordinates for downscale prediction.}
}
\value{
An object of class \code{sptPointPredictTimeAgg} containing the posterior predictive samples.
}
\description{
Point-level spatial-temporal prediction
}
