% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmCOS.R
\name{waicLM}
\alias{waicLM}
\title{Find ELPD for a fitted model using WAIC}
\usage{
waicLM(fit)
}
\arguments{
\item{fit}{An object of class \code{fitLM} or \code{modularLM} containing the fitted model.}
}
\value{
A numeric value representing the WAIC (Widely Applicable Information Criterion) for the fitted model.
}
\description{
Find ELPD for a fitted model using WAIC
}
